refactor: complete type migration from @types to standardized structure
- Removed src/core/@types/ directory entirely
- Updated 31 files to use new type import paths
  - 18 component files updated
  - 13 test files updated
- Added missing isInspectableComponent helper to components.ts
- Updated type-hierarchy.md to reflect completed migration
- Updated standardization-progress.md:
  - Marked Component Update Patterns as completed
  - Added Type Migration Completion section
  - Updated version to 4.20.0
- All tests passing with new structure

This completes the type standardization effort started in v4.18.4.
