+++ docs/standardization-progress.md
@@ -207,27 +207,39 @@ if (validation.valid) {
   - Proper component initialization after reset
   - Migration from legacy save format with component-level delegation
 
-### 8. 📋 Component Update Patterns
-**Standardize:**
-- Refresh rates (create constants)
-- Update hooks (unified pattern)
-- Debug update intervals
+### 8. ✅ Component Update Patterns (v4.19.0)
+**Status:** Completed
 
-## Next Session Priorities
+**Created:** `src/constants/ui.ts`
 
-1. **Complete Type Migration** (High Priority)
-   - Start with apple1/types/
-   - Update remaining core imports
-   - Clean up old directories
+**Key Constants:**
+- `DEBUG_REFRESH_RATES` - Standardized refresh rates for debug components
+- `UI_TIMINGS` - Animation and effect timings
+- `UPDATE_PATTERNS` - Component update strategies
 
-2. **Continue Formatter Migration** (Medium Priority)
-   - Focus on high-instance files
-   - Maintain consistency
+**Implementation:**
+```typescript
+export const DEBUG_REFRESH_RATES = {
+  CPU_STATUS: 100,     // CPU register updates
+  MEMORY_VIEW: 500,    // Memory viewer refresh
+  DISASSEMBLY: 200,    // Disassembler updates
+  STACK_VIEW: 100,     // Stack viewer refresh
+  INSPECTOR: 1000,     // Component inspector
+  CLOCK_DATA: 100      // Clock statistics
+} as const;
+```
 
-3. **Document Architecture** (Medium Priority)
-   - Create architecture diagram
-   - Document type hierarchy
-   - Update CLAUDE.md
+## Future Improvements
+
+1. **Documentation Cleanup** (Medium Priority)
+   - Organize docs into active/archive folders
+   - Update architecture documentation
+   - Remove redundant analysis files
+
+2. **Performance Optimization** (Low Priority)
+   - Investigate Bus cache performance
+   - Profile debug component updates
+   - Optimize Worker message handling
 
 ## Code Quality Metrics
 
@@ -271,5 +283,21 @@ grep -r "from.*@types/" src --include="*.ts" --include="*.tsx"
 ```
 
 ---
+### 9. ✅ Type Migration Completion (v4.20.0)
+**Status:** Completed on 2025-01-08
+
+**Migration Summary:**
+- ✅ Removed `src/core/@types/` directory entirely
+- ✅ Updated all imports (18 component files, 13 test files)
+- ✅ Added missing `isInspectableComponent` helper to `components.ts`
+- ✅ All tests passing with new structure
+
+**Key Changes:**
+- Migrated IInspectableComponent imports from `@types/` to `types/components`
+- Migrated BusSpaceType imports from `@types/IoAddressable` to `types/bus`
+- Removed InspectableArchView (unused, replaced by ArchViewNode)
+- Completed full type consolidation as outlined in type-hierarchy.md
+
+---
 Last updated: 2025-01-08
-Current version: 4.19.0
+Current version: 4.20.0
