--- src/core/@types/InspectableArchView.ts
@@ -1,6 +1,0 @@
-export interface InspectableArchView {
-    id: string;
-    type: string;
-    name?: string;
-    [key: string]: unknown; // Allow extension for component-specific fields
-}
