--- src/core/@types/IoComponent.ts
@@ -1,20 +1,0 @@
-export declare interface WireOptions {
-    write?: (value: number) => Promise<number | string | void>;
-    read?: (address: number) => Promise<number | void>;
-    reset?: () => void;
-}
-
-/**
- * Generic state type for IoComponents that support state save/restore
- */
-export type IoComponentState = Record<string, unknown>;
-
-export declare interface IoComponent<TState = IoComponentState> {
-    read?(address: number): Promise<number | void>; // Made optional
-    write(value: number | string): Promise<number | string | void>;
-    wire(options: WireOptions): void;
-    reset(): void;
-    // Optional: for components that support state save/restore
-    getState?(): TState;
-    setState?(state: TState): void;
-}
