+++ src/core/types/components.ts
@@ -110,3 +110,17 @@ export interface ArchitectureView {
         type: 'data' | 'control' | 'clock';
     }>;
 }
+
+/**
+ * Helper type guard for runtime checking
+ */
+export function isInspectableComponent(obj: unknown): obj is IInspectableComponent {
+    return (
+        typeof obj === 'object' &&
+        obj !== null &&
+        'id' in obj &&
+        typeof (obj as { id: unknown }).id === 'string' &&
+        'type' in obj &&
+        typeof (obj as { type: unknown }).type === 'string'
+    );
+}
