+++ src/version.ts
@@ -1,1 +1,1 @@
-export const APP_VERSION = '4.20.0';
+export const APP_VERSION = '4.21.0';
